+++ Forge/src/main/java/fuzs/armorstatues/client/ArmorStatuesForgeClient.java
@@ -4,8 +4,10 @@ import fuzs.armorstatues.ArmorStatues;
 import fuzs.armorstatues.api.ArmorStatuesApi;
 import fuzs.armorstatues.api.client.ArmorStatuesApiClient;
 import fuzs.armorstatues.client.handler.ArmorStandTooltipHandler;
+import fuzs.armorstatues.handler.ArmorStandInteractHandler;
 import fuzs.puzzleslib.client.core.ClientCoreServices;
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.ClientPlayerNetworkEvent;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -25,6 +27,9 @@ public class ArmorStatuesForgeClient {
     private static void registerHandlers() {
         MinecraftForge.EVENT_BUS.addListener((final ItemTooltipEvent evt) -> {
             ArmorStandTooltipHandler.onItemTooltip(evt.getItemStack(), evt.getFlags(), evt.getToolTip());
+        });
+        MinecraftForge.EVENT_BUS.addListener((final ClientPlayerNetworkEvent.LoggingIn evt) -> {
+            ArmorStandInteractHandler.clearPresentServerside();
         });
     }
 }