+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.0.1-1.19.2] - 2022-10-15
+### Changed
+- Updated to Forge 43.1.40+ which is also now required
+### Fixed
+- Armor stand interactions are now properly disabled when trying to open the menu when the mod is only installed client-side
+
 ## [v4.0.0-1.19.2] - 2022-09-22
 - Initial release